Create hook useMedicatioGroup
